moving http handler to common
