+++ libs/api/src/routing/index.ts
@@ -1,3 +1,2 @@
 export { default as createVideoRouter } from './videoRouter';
-export { default as httpHandler } from './httpHandler';
 export { default as createVideoHandler } from './videoHandler';
