+++ libs/api/src/types/index.ts
@@ -19,14 +19,16 @@ export type {
 } from '../schemas';
 
 export type * from './IVideoRouter';
-export type * from './ApplicationErrorMiddleware';
-export type * from './ApplicationHandler';
-export type * from './ApplicationRequest';
-export type * from './ApplicationRequestHandler';
-export type * from './ParamsDictionary';
-export type * from './Query';
+
 export * from './TokenRole';
 export * from './VideoAction';
 export type * from './HandlerConfig';
 export type * from './HandlersConfig';
 export type * from './HandlersDefaults';
+
+export type * from '@node/types/ApplicationErrorMiddleware';
+export type * from '@node/types/ApplicationHandler';
+export type * from '@node/types/ApplicationRequest';
+export type * from '@node/types/ApplicationRequestHandler';
+export type * from '@node/types/ParamsDictionary';
+export type * from '@node/types/Query';
