+++ libs/common/manifest.json
@@ -1,5 +1,5 @@
 {
   "version": "0.0.1",
   "channel": "latest",
-  "hash": "03f474273cc43f9e49444d5de0e6a922a4ae26604d81b69744071966ebdf3f0e"
+  "hash": "9d5e2e0839c4f95370e945fa9035fbbdb6e7a7ad3a33c520172c301613ba64bd"
 }
