+++ libs/common/srcNode/index.ts
@@ -2,5 +2,6 @@ export * from '../src';
 
 export * from './assertions';
 export * from './helpers';
+export * from './routing';
 export * from './schemas';
 export * from './types';
