+++ libs/common/srcNode/routing/index.ts
@@ -1,0 +1,1 @@
+export { default as httpHandler } from './httpHandler';
