+++ libs/common/srcNode/types/Query.ts
@@ -1,5 +1,5 @@
-export interface Query {
+export type Query = {
   [key: string]: undefined | string | Query | (string | Query)[];
-}
+};
 
 export default Query;
