+++ libs/common/srcNode/types/index.ts
@@ -1,5 +1,12 @@
 export * from '../../src/types';
 
+export type * from './ApplicationErrorMiddleware';
+export type * from './ApplicationHandler';
+export type * from './ApplicationRequest';
+export type * from './ApplicationRequestHandler';
+export type * from './ParamsDictionary';
+export type * from './Query';
+
 export type { TranscriptionProperties } from '../schemas/TranscriptionProperties.schema';
 export type { BaseArchiveOptions } from '../schemas/BaseArchiveOptions.schema';
 export type { ArchiveOptionsWithMaxBitrate } from '../schemas/ArchiveOptionsWithMaxBitrate.schema';
